LPS-54071 add interface for registry
