+++ portal-impl/src/META-INF/management-spring.xml
@@ -8,7 +8,7 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
 >
 	<alias name="mBeanServer" alias="mbeanServer" />
-	<bean id="mBeanRegistry" class="com.liferay.portal.kernel.jmx.MBeanRegistry">
+	<bean id="mBeanRegistry" class="com.liferay.portal.kernel.jmx.MBeanRegistryImpl">
 		 <property name="MBeanServer" ref="mBeanServer" />
 	</bean>
 	<bean id="mBeanServer" class="org.springframework.jmx.support.MBeanServerFactoryBean">
